% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{famos.run}
\alias{famos.run}
\title{FAMoS performance test run}
\format{A matrix with 7 rows and 17 columns}
\source{
Created by \code{\link{famos}}.
}
\usage{
famos.run
}
\description{
The data shows the performance of the FAMoS over the course of 8 iterations.
}
\keyword{datasets}
