% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_directories_function.R
\name{make.directories}
\alias{make.directories}
\title{Create Results Directories}
\usage{
make.directories(homedir)
}
\arguments{
\item{homedir}{The directory in which the result folders will be created.}
}
\value{
Creates directories.
}
\description{
Creates the directories, in which the results are going to be saved.
}
\details{
All files will be stored in the main folder "FAMoS-Results". It contains the following subdirectories:
\describe{
  \item{BestModel}{Contains the information criteria and the corresponding parameter estimates of the best fitting model.}
  \item{Figures}{Contains figures showing the performance of the FAMoS.}
  \item{Fits}{Contains the fitted parameter values of each of the tested models.}
  \item{TestedModels}{Contains the binary information of all of the tested models.}
  }
}
